requirements.txt
click==6.7
dask==0.19.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PySastrawi==1.2.0
six==1.11.0
Werkzeug==0.14.1
